Added Hammer
I kinda goofed and did the Table instead of the hammer. So, here's the hammer.

I am organizing the projects. Because of the goof, I am going to place Lab 01a and challenge 01a in the same folder. I'm just not sure what trying to move projects to new folders would do after I already set the project to it.