Chall1 and Lab 1
Okay, So this folder should have everything for the Chall 1 and Lab 1 assignments. .Ma exports and Screenshots for each: Hammer, Chair, Table, Lamp, Personal Project.

NOTE: I'd like to touch up the torch. I think for this project it's pretty good. I didn't have the extra whisps of the fire wrap all the way around because the game I was designing it for will be 2.5. 3D objects but 2D gameplay, so the player won't be seeing the backside and it would make it inefficient to model it. I tried to have like splits down the length of the wood, maybe that is something that Surfacing will handle, but you can't really notice the indent on the model. I'd also like to fix the wrap when we get that better figured out. For the rope at the bottom, I didn't find a better way to line it up than just eyeballing and moving sets of edges/vertices. So, I'm sure that could be done better.