Added Most Challenge and Lab b
Added: Hammer, Table, lamp for Lab and Challenge 1b.
Note: I went basic with the Hammer like instructed with the video. I did a good job, I've tried to be more creative, but I wasn't sure what to do for this one, and decided the time save was worth it.