Added Substance files for Chair and hammer
There should be a maps export, a painted export, and a screen shot for both chair and hammer.

There might also be fixed UVs for the Hammer, but I don't know if it saved that when I exported the new UV maps to use in substance painter.

The hammer model was a bit simple, and I was a little self-conscious that I was so lazy with the design. But I think I did enough with painting. The leather handle is rubbing off, and the rest of the hammer is rusted with some light dust cover. Not groundbreaking, but enough for the class I expect.