Added Surfacing challenge
The surfacing folder should have everything for the: Chair, Hammer, Table, Lamp, and make your own.

The torch didn't turn out quite as well as I wanted for two reasons.  I used an emissive layer to make the fire shine. The issues are that it caused the colors to all brighten significantly to have the glow be apparent, and even though I had to add an emmissive channel for many different layers they all had the same settings. So darker colors glowed less than lighter ones and I couldn't compensate by increasing the emmissive setting on only the dimmer matts. Still, all turned out well I think.

Also, I had some trouble with the table because I couldn't figure out how to have the effects not be white. I used a dust and a water droplet effect and they both painted on white and I couldn't figure out where to change the colors for that. It all still looks good, and I presume are well enough for the assignment, but that'll be stuff I'll look into.